Enable topological sort of the module's globla entities, i.e. types/consts/etc.
In general this is needed to ensure proper ordering which was previously ensured by using MapVector
and hoping everything's added to it in the correct order (not always true).

Currently it breaks up to 20 LITs due to the unstable (correct from the spec PoV though) types/consts declaration order.
Also conecptually fixes transcoding/extract_insert_value.ll which had
a wrong ordering before which wasn't covered with LIT checks